Implement comment addition to reports
The code now includes the necessary functionality to add comments to user reports, allowing for more detailed moderation. Also, a new "reports" field has been added to the User model, which counts the number of reports associated with a user. Minor fixes have also been provided to the User and Report models.